+++ CHANGELOG.md
@@ -488,3 +488,5 @@ All notable changes to Project Genesis will be documented in this file.
 
 ## Next Steps
 - [x] Implemented `is_deep_water_dancer` trait. Entities with this trait gain energy when on `deep-water` terrain.
+
+- Implemented `is_lava_dancer` trait. Entities with this trait gain energy when on `lava` terrain.